jenkins/custom_tools/DevopsJsonConv/src/main/java/com/infosys/json/Protractor.java
/***********************************************************************************************
*
* Copyright 2018 Infosys Ltd.
* Use of this source code is governed by MIT license that can be found in the LICENSE file or at
* https://opensource.org/licenses/MIT.
*
***********************************************************************************************/
package com.infosys.json;

import com.google.gson.annotations.Expose;
import com.google.gson.annotations.SerializedName;

public class Protractor {
	@SerializedName("totalTest")
	@Expose
	private Integer totalTest;
	@SerializedName("pass")
	@Expose
	private Integer pass;
	@SerializedName("fail")
	@Expose
	private Integer fail;
	@SerializedName("skipped")
	@Expose
	private Integer skipped;
	@SerializedName("errors")
	@Expose
	private Integer errors;

	public Integer getFail() {
		return fail;
	}

	public void setFail(Integer fail) {
		this.fail = fail;
	}

	public Integer getSkipped() {
		return skipped;
	}

	public void setSkipped(Integer skipped) {
		this.skipped = skipped;
	}

	public Integer getTotalTest() {
		return totalTest;
	}

	public void setTotaltest(Integer totalTest) {
		this.totalTest = totalTest;
	}

	public Integer getErrors() {
		return errors;
	}

	public void setErrors(Integer errors) {
		this.errors = errors;
	}

	public Integer getPass() {
		return pass;
	}

	public void setPass(Integer pass) {
		this.pass = pass;
	}
}
